refactor: extract shared common utilities to lib/common.ts
