+++ client-components/packages/camunda-api-zod-schemas/lib/8.8/ad-hoc-sub-process.ts
@@ -7,7 +7,7 @@
  */
 
 import {z} from 'zod';
-import {API_VERSION, getCollectionResponseBodySchema, type Endpoint} from './common';
+import {API_VERSION, getCollectionResponseBodySchema, type Endpoint} from '../common';
 
 const activityTypeSchema = z.enum([
 	'UNSPECIFIED',
