+++ client-components/packages/camunda-api-zod-schemas/lib/8.8/incident.ts
@@ -14,7 +14,7 @@ import {
 	advancedDateTimeFilterSchema,
 	getEnumFilterSchema,
 	type Endpoint,
-} from './common';
+} from '../common';
 
 const incidentErrorTypeSchema = z.enum([
 	'UNSPECIFIED',
