+++ client-components/packages/camunda-api-zod-schemas/lib/8.8/index.ts
@@ -667,7 +667,7 @@ export {
 	type QueryPage,
 	type QuerySortOrder,
 	type ProblemDetailsResponse,
-} from './common';
+} from '../common';
 export {
 	createDeploymentResponseBodySchema,
 	deleteResourceRequestBodySchema,
