+++ client-components/packages/camunda-api-zod-schemas/lib/8.8/job.ts
@@ -15,7 +15,7 @@ import {
 	getEnumFilterSchema,
 	type Endpoint,
 	basicStringFilterSchema,
-} from './common';
+} from '../common';
 
 const jobStateSchema = z.enum([
 	'CREATED',
