+++ client-components/packages/camunda-api-zod-schemas/lib/8.8/mapping-rule.ts
@@ -7,7 +7,7 @@
  */
 
 import {z} from 'zod';
-import {API_VERSION, getQueryRequestBodySchema, getQueryResponseBodySchema, type Endpoint} from './common';
+import {API_VERSION, getQueryRequestBodySchema, getQueryResponseBodySchema, type Endpoint} from '../common';
 
 const mappingRuleSchema = z.object({
 	mappingId: z.string(),
