+++ client-components/packages/camunda-api-zod-schemas/lib/8.8/message-subscriptions.ts
@@ -16,7 +16,7 @@ import {
 	getEnumFilterSchema,
 	getQueryRequestBodySchema,
 	getQueryResponseBodySchema,
-} from './common';
+} from '../common';
 
 const messageSubscriptionStateSchema = z.enum(['CORRELATED', 'CREATED', 'DELETED', 'MIGRATED']);
 type MessageSubscriptionState = z.infer<typeof messageSubscriptionStateSchema>;
