+++ client-components/packages/camunda-api-zod-schemas/lib/8.8/process-definition.ts
@@ -18,7 +18,7 @@ import {
 	basicStringFilterSchema,
 	getOrFilterSchema,
 	advancedIntegerFilterSchema,
-} from './common';
+} from '../common';
 import {
 	processDefinitionSchema,
 	processDefinitionStatisticSchema,
