+++ client-components/packages/camunda-api-zod-schemas/lib/8.8/process-instance.ts
@@ -18,7 +18,7 @@ import {
 	basicStringFilterSchema,
 	getOrFilterSchema,
 	type Endpoint,
-} from './common';
+} from '../common';
 import {variableSchema} from './variable';
 import {
 	processInstanceSchema,
