+++ client-components/packages/camunda-api-zod-schemas/lib/8.8/tenant.ts
@@ -7,7 +7,7 @@
  */
 
 import {z} from 'zod';
-import {API_VERSION, getQueryRequestBodySchema, getQueryResponseBodySchema, type Endpoint} from './common';
+import {API_VERSION, getQueryRequestBodySchema, getQueryResponseBodySchema, type Endpoint} from '../common';
 import {queryGroupsRequestBodySchema, queryGroupsResponseBodySchema, type Group} from './group';
 import {queryRolesRequestBodySchema, queryRolesResponseBodySchema, type Role} from './role';
 import {
