+++ client-components/packages/camunda-api-zod-schemas/lib/8.8/user-task.ts
@@ -16,7 +16,7 @@ import {
 	getQueryRequestBodySchema,
 	getQueryResponseBodySchema,
 	type Endpoint,
-} from './common';
+} from '../common';
 import {variableSchema} from './variable';
 
 const userTaskVariableFilterSchema = variableSchema.pick({
